Fix challenge problems for the balanced words problem.
Messy but it works. The issue was with the calculation of the scores.

The problem came about when calculating the left hand side.
It is defined that the score for an index is:

Distance from Pivot * Space in alphabet

This became a problme when:

SOMETHING @ pivot = 2

SO M THING
21   12345

This is the word and it's distances below it. What we generated was:

SOMETHING @ pivot = 2

SO M THING
12   12345